SOC-1241 | Renaming mixins, adding JSDoc and model check to create(post)
