+++ front/main/app/mixins/discussion-forum-actions-controller.js
