+++ front/main/app/routes/discussion/reported-posts.js
@@ -3,14 +3,14 @@ import DiscussionRouteUpvoteMixin from '../../mixins/discussion-route-upvote';
 import DiscussionReportedPostsModel from '../../models/discussion-reported-posts';
 import DiscussionLayoutMixin from '../../mixins/discussion-layout';
 import DiscussionModerationRouteMixin from '../../mixins/discussion-moderation-route';
-import DiscussionFilteringRouteMixin from '../../mixins/discussion-filtering-route';
+import DiscussionForumActionsRouteMixin from '../../mixins/discussion-forum-actions-route';
 
 
 export default DiscussionBaseRoute.extend(
 	DiscussionLayoutMixin,
 	DiscussionRouteUpvoteMixin,
 	DiscussionModerationRouteMixin,
-	DiscussionFilteringRouteMixin,
+	DiscussionForumActionsRouteMixin,
 	{
 		discussionEditor: Ember.inject.service(),
 		discussionSort: Ember.inject.service(),
@@ -53,6 +53,13 @@ export default DiscussionBaseRoute.extend(
 				this.modelFor('discussion.reported-posts').loadPage(pageNum, this.get('discussionSort.sortBy'));
 			},
 
+			/**
+			 * Goes to post list page and attempts to create a new post there
+			 *
+			 * @param {object} postData
+			 *
+			 * @returns {void}
+			 */
 			create(postData) {
 				this.get('discussionSort').setSortBy('latest');
 				this.transitionTo('discussion.forum', this.get('forumId'), 'latest').promise.then(() => {
